re-configure navigation stack/tab
